% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drif.R
\name{drif}
\alias{drif}
\title{Decomposite Recentered Influence Function regression}
\usage{
drif()
}
\arguments{
\item{object}{a recentered influence function regression model}

\item{tau}{quantile}
}
\description{
Decomposite Recentered Influence Function regression
}

